fix: support using existing dns_zone
This change introduces a flag to use existing DNS Zone when it already
exists, preventing errors on terraform when the user specified an domain
in BaseDomain (Assisted Installer) which has been created, or is already
used by the company.

The script will follow the same flow, but reuse existing zone to
create cluster DNS records.