Customize network configuration to install OCP 4.17
From OCP 4.17 the network operator will use the range 169.254.0.0/17 for
the `internalMasqueradeSubnet` (https://docs.openshift.com/container-platform/4.16/networking/cluster-network-operator.html#nw-operator-cr-cno-object_cluster-network-operator).

This default value collides with the iSCSI subnet used in OCI, and
prevent the installation of OCP on nodes with iSCSI voot drives.

This commit updates the default value of `internalMasqueradeSubnet` to
use a range that dosn´t not collide to anything [known to me] in Oracle
Cloud.